# Conflicts:
#	BuildSystem/Version.iss
#	BuildSystem/buildtools/classes/Archiver.py
#	BuildSystem/buildtools/classes/DebBuilder.py
#	CASAL2/source/Version.h
#	Documentation/UserManual/Version.tex
#	R-libraries/Version.R
#	R-libraries/casal2/DESCRIPTION
#	R-libraries/casal2/R/Casal2.binary.version.R
#	R-libraries/casal2/R/Version.R
#	TestCases/primary/BCO/Casal2/adolc_casal_flags_off/run_estimation.txt
#	TestCases/primary/BCO/Casal2/adolc_casal_flags_on/run_estimation.txt
#	TestCases/primary/BCO/Casal2/adolc_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/BCO/Casal2/betadiff_casal_flags_off/run_estimation.txt
#	TestCases/primary/BCO/Casal2/betadiff_casal_flags_on/run_estimation.txt
#	TestCases/primary/BCO/Casal2/betadiff_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/BCO/compare_BCO.MPD.html
#	TestCases/primary/HAK/Casal2/adolc_casal_flags_off/run_estimation.txt
#	TestCases/primary/HAK/Casal2/adolc_casal_flags_on/run_estimation.txt
#	TestCases/primary/HAK/Casal2/adolc_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/HAK/Casal2/betadiff_casal_flags_off/run_estimation.txt
#	TestCases/primary/HAK/Casal2/betadiff_casal_flags_on/run_estimation.txt
#	TestCases/primary/HAK/Casal2/betadiff_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/HAK/compare_HAK.MPD.html
#	TestCases/primary/HOK/Casal2/adolc_casal_flags_off/run_estimation.txt
#	TestCases/primary/HOK/Casal2/adolc_casal_flags_on/run_estimation.txt
#	TestCases/primary/HOK/Casal2/adolc_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/HOK/Casal2/betadiff_casal_flags_off/run_estimation.txt
#	TestCases/primary/HOK/Casal2/betadiff_casal_flags_on/run_estimation.txt
#	TestCases/primary/HOK/Casal2/betadiff_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/HOK/compare_HOK.MPD.html
#	TestCases/primary/LIN/Casal2/adolc_casal_flags_off/run_estimation.txt
#	TestCases/primary/LIN/Casal2/adolc_casal_flags_on/run_estimation.txt
#	TestCases/primary/LIN/Casal2/adolc_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/LIN/Casal2/betadiff_casal_flags_off/run_estimation.txt
#	TestCases/primary/LIN/Casal2/betadiff_casal_flags_on/run_estimation.txt
#	TestCases/primary/LIN/Casal2/betadiff_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/LIN/compare_LIN.MPD.html
#	TestCases/primary/ORH/Casal2/adolc_casal_flags_off/run_estimation.txt
#	TestCases/primary/ORH/Casal2/adolc_casal_flags_on/run_estimation.txt
#	TestCases/primary/ORH/Casal2/adolc_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/ORH/Casal2/betadiff_casal_flags_off/run_estimation.txt
#	TestCases/primary/ORH/Casal2/betadiff_casal_flags_on/run_estimation.txt
#	TestCases/primary/ORH/Casal2/betadiff_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/ORH/compare_ORH.MPD.html
#	TestCases/primary/SBW/Casal2/adolc_casal_flags_off/run_estimation.txt
#	TestCases/primary/SBW/Casal2/adolc_casal_flags_on/run_estimation.txt
#	TestCases/primary/SBW/Casal2/adolc_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/SBW/Casal2/betadiff_casal_flags_off/run_estimation.txt
#	TestCases/primary/SBW/Casal2/betadiff_casal_flags_on/run_estimation.txt
#	TestCases/primary/SBW/Casal2/betadiff_casal_flags_on_low_tol/run_estimation.txt
#	TestCases/primary/SBW/compare_SBW.MPD.html
#	TestModels/Complex/run.log
#	TestModels/Complex_input/run.log
#	TestModels/SBW/estimate.log
#	TestModels/Simple/estimate.log
#	TestModels/Simple/log.out
#	TestModels/SingleSexTagByLength_input/report.csl2
#	TestModels/SingleSexTagByLength_input/run.log
#	TestModels/TwoSex/estimate_betadiff.log
#	TestModels/TwoSex_input/run.log
#	TestModels/agelength_cv_with_LF/run.log

+++ CASAL2/source/AgeingErrors/Age/Normal.cpp
@@ -82,7 +82,7 @@ void Normal::DoBuild() {
  * changes from any addressable modifications
  */
 void Normal::DoReset() {
-  Double age           = 0.0;
+  unsigned age         = 0;
   Double min_age_class = 0.0;
 
   for (unsigned i = 0; i < age_spread_; ++i) {
